同步

+++ gw_full_latest/cmd.sh
@@ -26,23 +26,37 @@ choice="Z"
 if [ ! -f gateway_id.txt ]
 then
 	echo "ERROR: gateway_id.txt file not found"
-	echo "should create it by running echo \"000000XXXXXXXXXX\" > gateway_id.txt"
-	echo "where XXXXXXXXXX is the last 5 bytes of your MAC Ethernet interface address"
-	echo "Example: echo \"00000027EBBEDA21\" > gateway_id.txt"
-	echo "Here is your MAC Ethernet interface address:"
-	echo "-------------------------------------------------------"
-	ifconfig | grep "eth0"
-    echo "-------------------------------------------------------"
-	echo "Enter the last 5 hex bytes of your MAC Ethernet interface address"
-	echo "in capital character and without the : separator"
-	echo "example: HWaddr b8:27:eb:be:da:21 then just enter 27EBBEDA21"
-	read macaddr
-	echo "Will write 000000$macaddr into gateway_id.txt"
-	echo "000000$macaddr" > gateway_id.txt
+	echo "Reading MAC address to get last 5 bytes for gateway id"
+	
+	#get the last 5 bytes of the eth0 MAC addr
+	gwid=`ifconfig | grep 'eth0' | awk '{print $NF}' | sed 's/://g' | awk '{ print toupper($1) }' | cut -c 3-`
+
+	#get the last 5 bytes of the wlan0 MAC addr
+	if [ "$gwid" = "" ]
+		then
+			gwid=`ifconfig | grep 'wlan0' | awk '{print $NF}' | sed 's/://g' | awk '{ print toupper($1) }' | cut -c 3-`
+			
+			#it means that the wlan0 interface works in access point mode or has no IP address assigned
+			#so get the address from the ether field
+			if [ "$gwid" = "00" ]
+			then
+				gwid=`ifconfig | grep 'ether' | awk '{print $2}' | sed 's/://g' | awk '{ print toupper($1) }' | cut -c 3-`
+			fi
+	fi
+
+	#set a default value
+	if [ "$gwid" = "" ]
+		then
+			gwid="XXXXXXDEF0"
+	fi
+	
+	echo "Creating gateway_id.txt file"
+	echo "Writing 000000$gwid"
+	echo "000000$gwid" > gateway_id.txt
 	echo "Done"
 	echo "Replacing gw id in gateway_conf.json"
-	sed -i -- 's/"000000.*"/"000000'"$macaddr"'"/g' gateway_conf.json
-	echo "Done"	
+	sed -i -- 's/"000000.*"/"000000'"$gwid"'"/g' gateway_conf.json
+	echo "Done"
 fi
 
 gatewayid=`cat gateway_id.txt`
@@ -50,13 +64,13 @@ gatewayid=`cat gateway_id.txt`
 while [ "$choice" != "Q" ]
 do
 echo "=======================================* Gateway $gatewayid *==="
-echo "0- sudo python start_gw.py & ; disown %1                             +"
+echo "0- sudo python start_gw.py                                           +"
 echo "1- sudo ./lora_gateway --mode 1                                      +"
 echo "2- sudo ./lora_gateway --mode 1 | python post_processing_gw.py       +"
 echo "3- ps aux | grep -e start_gw -e lora_gateway -e post_proc -e log_gw  +"
-echo "4- tail --line=25 ../Dropbox/LoRa-test/post-processing_*.log         +"
-echo "5- tail --line=25 -f ../Dropbox/LoRa-test/post-processing_*.log      +"
-echo "6- less ../Dropbox/LoRa-test/post-processing_*.log                   +"
+echo "4- tail --line=25 ../Dropbox/LoRa-test/post-processing.log           +"
+echo "5- tail --line=25 -f ../Dropbox/LoRa-test/post-processing.log        +"
+echo "6- less ../Dropbox/LoRa-test/post-processing.log                     +"
 echo "------------------------------------------------------* Bluetooth *--+"
 echo "a- run: sudo hciconfig hci0 piscan                                   +"
 echo "b- run: sudo python rfcomm-server.py                                 +"
@@ -95,10 +109,10 @@ echo "BEGIN OUTPUT"
 
 if [ "$choice" = "0" ]
 	then
-		echo "Running in background the full LoRa gateway" 
-		sudo python start_gw.py &
-		disown %1
-		echo "Check ../Dropbox/LoRa-test/post-processing_$gatewayid.log file (select command 5)"
+		echo "Running interactively the full LoRa gateway... CTRL-C to exit" 
+		sudo python start_gw.py
+		#disown %1
+		#echo "Check ../Dropbox/LoRa-test/post-processing_$gatewayid.log file (select command 5)"
 fi
 
 if [ "$choice" = "1" ] 
@@ -125,28 +139,28 @@ fi
 
 if [ "$choice" = "4" ] 
 	then
-		echo "Displaying last 25 lines of ../Dropbox/LoRa-test/post-processing_$gatewayid.log"
+		echo "Displaying last 25 lines of ../Dropbox/LoRa-test/post-processing.log"
 		echo "Current UTC date is"
 		date --utc
-		tail --line=25 ../Dropbox/LoRa-test/post-processing_$gatewayid.log
+		tail --line=25 ../Dropbox/LoRa-test/post-processing.log
 fi
 
 if [ "$choice" = "5" ] 
 	then
-		echo "Following last lines of ../Dropbox/LoRa-test/post-processing_$gatewayid.log. CTRL-C to return"
+		echo "Following last lines of ../Dropbox/LoRa-test/post-processing.log. CTRL-C to return"
 		echo "Current UTC date is"
 		date --utc
 		trap "echo" SIGINT
-		tail --line=25 -f ../Dropbox/LoRa-test/post-processing_$gatewayid.log
+		tail --line=25 -f ../Dropbox/LoRa-test/post-processing.log
 fi
 
 if [ "$choice" = "6" ] 
 	then
-		echo "Display ../Dropbox/LoRa-test/post-processing_$gatewayid.log. Q to return"
+		echo "Display ../Dropbox/LoRa-test/post-processing.log. Q to return"
 		echo "Current UTC date is"
 		date --utc
 		trap "echo" SIGINT
-		less ../Dropbox/LoRa-test/post-processing_$gatewayid.log
+		less ../Dropbox/LoRa-test/post-processing.log
 fi
 
 if [ "$choice" = "a" ] 
@@ -199,91 +213,91 @@ fi
 if [ "$choice" = "l" ] 
 	then
 		echo "List LoRa reception indications"
-		grep -a "rxlora" ../Dropbox/LoRa-test/post-processing_$gatewayid.log
+		grep -a "rxlora" ../Dropbox/LoRa-test/post-processing.log
 fi
 
 if [ "$choice" = "m" ] 
 	then
 		echo "List radio module reset"
-		grep -a "Resetting radio module" ../Dropbox/LoRa-test/post-processing_$gatewayid.log
+		grep -a "Resetting radio module" ../Dropbox/LoRa-test/post-processing.log
 fi
 
 if [ "$choice" = "n" ] 
 	then
 		echo "List boot indications"
-		grep -a "**********Power ON" ../Dropbox/LoRa-test/post-processing_$gatewayid.log
+		grep -a "**********Power ON" ../Dropbox/LoRa-test/post-processing.log
 fi
 
 if [ "$choice" = "o" ] 
 	then
 		echo "List post-processing status"
-		grep -a "post status: gw ON" ../Dropbox/LoRa-test/post-processing_$gatewayid.log
+		grep -a "post status: gw ON" ../Dropbox/LoRa-test/post-processing.log
 fi
 
 if [ "$choice" = "p" ] 
 	then
 		echo "List low-level gateway status"
-		grep -a "Low-level gw status ON" ../Dropbox/LoRa-test/post-processing_$gatewayid.log
+		grep -a "Low-level gw status ON" ../Dropbox/LoRa-test/post-processing.log
 fi
 
 if [ "$choice" = "A" ] 
-        then
-                echo "Showing global_conf.json"
-                cat gateway_conf.json
+    then
+    	echo "Showing global_conf.json"
+		cat gateway_conf.json
 fi
 
 if [ "$choice" = "B" ] 
-        then
-        		if [ -f gateway_conf.json ]
-        			then
-                		echo "Editing gateway_conf.json. CTRL-O to save, CTRL-X to return"
-                		nano gateway_conf.json
-                	else
-                		echo "Error: gateway_conf.json does not exist"
-                fi	
+    then
+    	if [ -f gateway_conf.json ]
+        	then
+            	echo "Editing gateway_conf.json. CTRL-O to save, CTRL-X to return"
+                nano gateway_conf.json
+            else
+                echo "Error: gateway_conf.json does not exist"
+        fi	
 fi
 
 if [ "$choice" = "C" ] 
-        then
-                echo "Showing clouds.json"
-                cat clouds.json
+    then
+    	echo "Showing clouds.json"
+        cat clouds.json
 fi
 
 if [ "$choice" = "D" ] 
-        then
-        		if [ -f clouds.json ]
-        			then
-                		echo "Editing clouds.json. CTRL-O to save, CTRL-X to return"
-                		nano clouds.json
-                	else
-                		echo "Error: clouds.json does not exist"
-                fi	
+    then
+        if [ -f clouds.json ]
+        	then
+                echo "Editing clouds.json. CTRL-O to save, CTRL-X to return"
+                nano clouds.json
+            else
+                echo "Error: clouds.json does not exist"
+        fi	
 fi
 
 if [ "$choice" = "U" ]
-		then
-			echo "Updating to latest version of gateway, preserving your local configuration file"
-			pushd scripts
-			./update_gw.sh
-			popd
+	then
+		echo "Updating to latest version of gateway, preserving your local configuration file"
+		cd /home/pi/lora_gateway/scripts
+		./update_gw.sh
+		cd /home/pi/lora_gateway
 fi
 
 if [ "$choice" = "V" ]
-		then
-			echo "Download and install a file"
-			echo "Enter the URL of the file:"
-			read filename_url			
-			wget --backups=1 $filename_url
-			echo "Done"
+	then
+		echo "Download and install a file"
+		echo "Enter the URL of the file:"
+		read filename_url			
+		wget --backups=1 $filename_url
+		echo "Done"
 fi
 
 if [ "$choice" = "W" ]
-		then
-			echo "Run a command"
-			echo "Enter the command to run:"
-			read command_line			
-			eval $command_line
-			echo "Done"
+	then
+		echo "Run a command"
+		echo "Enter the command to run:"
+		read command_line			
+		eval $command_line
+		echo "Done"
 fi
 
 
@@ -294,21 +308,37 @@ if [ "$choice" = "K" ]
 fi
 
 if [ "$choice" = "k" ] 
-        then
+    then
 		echo "Killing rfcomm-server process"
 		sudo kill $(ps aux | grep -e rfcomm-server | awk '{print $2}')
 fi
 
 if [ "$choice" = "R" ] 
-        then
-		echo "Reboot gateway"
-		sudo shutdown -r now
+    then
+		echo "*******************"
+		echo "*** reboot Y/N  ***"
+		echo "*******************"
+		read ouinon
+
+		if [ "$ouinon" = "y" ] || [ "$ouinon" = "Y" ]
+			then    
+				echo "Reboot gateway"
+				sudo shutdown -r now
+		fi		
 fi
 
 if [ "$choice" = "S" ] 
-        then
-		echo "Shutdown gateway"
-		sudo shutdown -h now
+    then
+		echo "*********************"
+		echo "*** shutdown Y/N  ***"
+		echo "*********************"
+		read ouinon
+
+		if [ "$ouinon" = "y" ] || [ "$ouinon" = "Y" ]
+			then    
+				echo "Shutdown gateway"
+				sudo shutdown -h now
+		fi		
 fi
 
 echo "END OUTPUT"
